refactor: Rename [Message -> ErrorMessage]
