+++ src/main/java/com/hexagonal/courseregistration/course/application/RegisterCourseUseCase.java
@@ -2,7 +2,7 @@ package com.hexagonal.courseregistration.course.application;
 
 import com.hexagonal.courseregistration.course.application.error.AuthorityException;
 import com.hexagonal.courseregistration.course.application.error.CourseException;
-import com.hexagonal.courseregistration.course.application.error.Message;
+import com.hexagonal.courseregistration.course.application.error.ErrorMessage;
 import com.hexagonal.courseregistration.course.application.port.CheckExistCoursePort;
 import com.hexagonal.courseregistration.course.application.port.CheckProfessorPort;
 import com.hexagonal.courseregistration.course.application.port.RegisterRequest;
@@ -17,11 +17,11 @@ public class RegisterCourseUseCase {
 
   public void register(RegisterRequest request) {
     if (!checkProfessorPort.check(request.userId())) {
-      throw new AuthorityException(Message.NOT_PROFESSOR);
+      throw new AuthorityException(ErrorMessage.NOT_PROFESSOR);
     }
 
     if (checkExistCoursePort.check(request.courseName())) {
-      throw new CourseException(Message.EXIST_COURSE);
+      throw new CourseException(ErrorMessage.EXIST_COURSE);
     }
 
     saveCoursePort.save(request);
