+++ src/main/java/com/hexagonal/courseregistration/course/application/error/AuthorityException.java
@@ -3,7 +3,7 @@ package com.hexagonal.courseregistration.course.application.error;
 import com.hexagonal.courseregistration.common.application.ForbiddenException;
 
 public class AuthorityException extends ForbiddenException {
-  public AuthorityException(Message message) {
-    super(message.detail());
+  public AuthorityException(ErrorMessage errorMessage) {
+    super(errorMessage.detail());
   }
 }
