+++ src/main/java/com/hexagonal/courseregistration/course/application/error/CourseException.java
@@ -3,7 +3,7 @@ package com.hexagonal.courseregistration.course.application.error;
 import com.hexagonal.courseregistration.common.application.DomainException;
 
 public class CourseException extends DomainException {
-  public CourseException(Message message) {
-    super(message.detail());
+  public CourseException(ErrorMessage errorMessage) {
+    super(errorMessage.detail());
   }
 }
