+++ src/main/java/com/hexagonal/courseregistration/course/application/error/UserException.java
@@ -3,7 +3,7 @@ package com.hexagonal.courseregistration.course.application.error;
 import com.hexagonal.courseregistration.common.application.DomainException;
 
 public class UserException extends DomainException {
-  public UserException(Message message) {
-    super(message.detail());
+  public UserException(ErrorMessage errorMessage) {
+    super(errorMessage.detail());
   }
 }
